backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
mysqlclient==2.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
pymysql==1.1.0
requests==2.31.0
cryptography==41.0.7
drf-spectacular==0.27.0
